requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7
django-all-access==0.6.0
django-cors-headers==0.12
django-tastypie==0.12.1
django-toolbelt==0.0.1
feedparser==5.1.3
gunicorn==18.0
jsonpickle==0.7.0
MechanicalSoup==0.2.0
oauthlib==0.6.1
Pillow==2.4.0
psycopg2==2.5.2
pycrypto==2.6.1
pystache==0.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.2.1
requests-oauthlib==0.4.0
simplejson==3.6.5
six==1.6.1
South==0.8.4
static==1.0.2
static3==0.5.1
